use existing extensions class.
