+++ src/Markup/Avalonia.Markup.Xaml/Extensions.cs
@@ -2,7 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
+using Avalonia.Controls;
 using Avalonia.Markup.Xaml.XamlIl.Runtime;
+using Avalonia.Styling;
 
 namespace Avalonia.Markup.Xaml
 {
@@ -31,6 +33,27 @@ namespace Avalonia.Markup.Xaml
             var tr = ctx.GetService<IXamlTypeResolver>();
             string name = string.IsNullOrEmpty(namespacePrefix) ? type : $"{namespacePrefix}:{type}";
             return tr?.Resolve(name);
+        }
+        
+        public static object GetDefaultAnchor(this IServiceProvider provider)
+        {
+            // If the target is not a control, so we need to find an anchor that will let us look
+            // up named controls and style resources. First look for the closest IControl in
+            // the context.
+            object anchor = provider.GetFirstParent<IControl>();
+
+            if (anchor is null)
+            {
+                // Try to find IDataContextProvider, this was added to allow us to find
+                // a datacontext for Application class when using NativeMenuItems.
+                anchor = provider.GetFirstParent<IDataContextProvider>();
+            }
+
+            // If a control was not found, then try to find the highest-level style as the XAML
+            // file could be a XAML file containing only styles.
+            return anchor ??
+                   provider.GetService<IRootObjectProvider>()?.RootObject as IStyle ??
+                   provider.GetLastParent<IStyle>();
         }
     }
 }
